sort demo snippets
